@@ -4,22 +4,45 @@
 
     <groupId>net.cubespace</groupId>
     <artifactId>geSuit-parent</artifactId>
-    <version>PARENT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>geSuit-parent</name>
 
     <properties>
-        <build.number>${BUILD_NUMBER}</build.number>
+        <build.number/>
+        <plugin.version>${project.version}-${build.number}</plugin.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
-
+    <developers>
+        <developer>
+            <id>AddstarMC</id>
+            <name>AddstarMC</name>
+            <url>http://github.com/AddstarMC</url>
+            <roles>
+                <role>developer</role>
+                <role>maintainer</role>
+            </roles>
+            <timezone>10</timezone>
+        </developer>
+    </developers>
+    <distributionManagement>
+        <repository>
+            <id>maven.addstar.com.au</id>
+            <name>maven.addstar.com.au-releases</name>
+            <url>http://maven.addstar.com.au/artifactory/ext-release-local</url>
+        </repository>
+        <snapshotRepository>
+            <id>maven.addstar.com.au</id>
+            <name>maven.addstar.com.au-snapshots</name>
+            <url>http://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
+        </snapshotRepository>
+    </distributionManagement>
     <modules>
         <module>bungee</module>
         <module>common</module>
-
         <!-- Bukkit Module Container -->
         <module>bukkit</module>
-
         <!-- geSuit Modules -->
         <module>modules/moderation</module>
         <module>modules/teleports</module>
@@ -37,10 +60,6 @@
                 <enabled>true</enabled>
             </snapshots>
         </repository>
-        <!--<repository>
-            <id>cubespace</id>
-            <url>http://nexus.cube-space.net/content/groups/public/</url>
-        </repository>-->
 	<repository>
 	    <id>addstar-repo</id>
 	    <url>http://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
@@ -48,6 +67,14 @@
 	      <enabled>true</enabled>
 	    </snapshots>
 	</repository>
+        <repository>
+            <id>maven.addstar.com.au</id>
+            <name>maven.addstar.com.au-releases</name>
+            <url>http://maven.addstar.com.au/artifactory/ext-release-local</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
 	<repository>
 	    <id>maven-central</id>
 	    <url>http://repo1.maven.org/maven2/</url>
@@ -71,13 +98,25 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.1</version>
                     <configuration>
-                        <source>1.7</source>
-                        <target>1.7</target>
+                        <source>1.8</source>
+                        <target>1.8</target>
                         <showWarnings>true</showWarnings>
                         <showDeprecation>true</showDeprecation>
                     </configuration>
                 </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-source-plugin</artifactId>
+                    <version>3.0.1</version>
+                </plugin>
             </plugins>
         </pluginManagement>
+        <plugins>
+            <plugin>
+                <artifactId>maven-release-plugin</artifactId>
+                <version>2.5.3</version>
+
+            </plugin>
+        </plugins>
     </build>
 </project>
